+++ android/src/main/java/com/serenegiant/usb/USBMonitor.java
@@ -167,7 +167,12 @@ public final class USBMonitor {
 			if (DEBUG) Log.i(TAG, "register:");
 			final Context context = mWeakContext.get();
 			if (context != null) {
-				mPermissionIntent = PendingIntent.getBroadcast(context, 0, new Intent(ACTION_USB_PERMISSION), 0);
+				// Android 12+ requires FLAG_IMMUTABLE or FLAG_MUTABLE
+				int flags = 0;
+				if (BuildCheck.isAPI31()) {
+					flags = PendingIntent.FLAG_MUTABLE;
+				}
+				mPermissionIntent = PendingIntent.getBroadcast(context, 0, new Intent(ACTION_USB_PERMISSION), flags);
 				final IntentFilter filter = new IntentFilter(ACTION_USB_PERMISSION);
 				// ACTION_USB_DEVICE_ATTACHED never comes on some devices so it should not be added here
 				filter.addAction(UsbManager.ACTION_USB_DEVICE_DETACHED);