+++ android/src/main/java/com/serenegiant/utils/BuildCheck.java
@@ -14,4 +14,8 @@ public class BuildCheck {
     public static boolean isAndroid5() {
         return Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP;
     }
+    
+    public static boolean isAPI31() {
+        return Build.VERSION.SDK_INT >= 31; // Build.VERSION_CODES.S
+    }
 }